chore: bump Rust toolchain to nightly-2026-08-28

+++ src/query/storages/common/index/src/hnsw_index/quantization/encoded_vectors_u8.rs
@@ -208,7 +208,7 @@ impl<TStorage: EncodedStorage> EncodedVectorsU8<TStorage> {
                     }
                 };
 
-                return self.metadata.multiplier * score as f32 + query.offset + vector_offset;
+                return self.metadata.multiplier * score + query.offset + vector_offset;
             }
         }
